backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
torch
torchvision
pillow
numpy
opencv-python
pandas
pymysql
python-dotenv